+++ src/coup/rep/mod.rs
@@ -31,83 +31,17 @@ use tracing::instrument;
 #[derive(Hash, PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Serialize, Deserialize)]
 pub struct Move(pub(crate) u16);
 
-#[rustfmt::skip] const SOURCE_IDX_MASK   : u16   = 0b111111_000000_0_000;
-#[rustfmt::skip] const SOURCE_IDX_SHIFT  : usize = 10;
-#[rustfmt::skip] const TARGET_IDX_MASK   : u16   = 0b000000_111111_0_000;
-#[rustfmt::skip] const TARGET_IDX_SHIFT  : usize = 4;
-#[rustfmt::skip] const METADATA_MASK     : u16   = 0b000000_000000_1_111;
-
+mod bitmangling;
+mod creation;
 mod debug;
 mod move_type;
 
+#[cfg(test)] mod tests;
+
 pub use move_type::*;
+pub use bitmangling::*;
 
 impl Move {
-    pub const fn empty() -> Move {
-        Move (0)
-    }
-
-    /// Creates a move from a given source and target index,
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::notation::*;
-    /// // the move from d2 -> d4
-    /// let m = Move::new(D2, D4, MoveType::QUIET);
-    /// assert_eq!(m.source(), D2);
-    /// assert_eq!(m.target(), D4);
-    /// assert!(!m.is_promotion());
-    /// assert!(m.move_metadata().is_quiet());
-    /// ```
-    pub fn new(source: impl Into<Square>, target: impl Into<Square>, metadata: MoveType) -> Move {
-        let s : Square = source.into();
-        let t : Square = target.into();
-
-        #[rustfmt::skip] Move( (s.index() as u16) << SOURCE_IDX_SHIFT
-                             | (t.index() as u16) << TARGET_IDX_SHIFT
-                             |   metadata as u16 )
-    }
-
-    pub fn set_metadata(&mut self, metadata: MoveType) {
-        self.0 = (self.0 & !METADATA_MASK) | (metadata as u16);
-    }
-
-    pub fn null() -> Move {
-        // We only care about the metadata bits for a null move. So the source/target are just
-        // whatever is convenient.
-        Move::new(A1, A1, MoveType::NULLMOVE)
-    }
-
-    /// Creates a move from the given source and target squares (given in notation), and
-    /// the provided metadata. If a Right(Piece) is provided, the move is assumed to be a
-    /// valid promotion. No error checking is done.
-    ///
-    /// NOTE: do not use this internally, this is for testing convenience!
-    ///
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::notation::*;
-    /// # use hazel::constants::*;
-    /// # use hazel::types::Piece;
-    /// // the move from d2 -> d4
-    /// let m = Move::from_notation("d2", "d4", MoveType::DOUBLE_PAWN);
-    ///
-    /// assert_eq!(m.source(), D2);
-    /// assert_eq!(m.target(), D4);
-    /// assert!(!m.is_promotion());
-    ///
-    /// let pm = Move::from_notation("d7", "d8", MoveType::PROMOTION_QUEEN);
-    /// assert_eq!(pm.source(), D7);
-    /// assert_eq!(pm.target(), D8);
-    /// assert!(pm.is_promotion());
-    /// assert_eq!(pm.promotion_piece(), Piece::Queen);
-    /// ```
-    pub fn from_notation(source: &str, target: &str, metadata: MoveType) -> Move {
-        Move::new(
-            Square::try_from(source).unwrap(),
-            Square::try_from(target).unwrap(),
-            metadata,
-        )
-    }
 
     /// Disambiguates the move in the context of the provided query. If the move is not marked ambiguous,
     /// the move is returned as is. If the move is ambiguous, the context is used to determine the
@@ -277,105 +211,7 @@ impl Move {
         }
     }
 
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::notation::*;
-    /// // the move from d2 -> d4
-    ///
-    /// let m = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-    /// assert_eq!(m.source(), D2);
-    /// ```
-    pub fn source_idx(&self) -> usize {
-        ((self.0 & SOURCE_IDX_MASK) >> SOURCE_IDX_SHIFT).into()
-    }
 
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::notation::*;
-    /// // the move from d2 -> d4
-    ///
-    /// let m = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-    /// assert_eq!(m.source(), D2);
-    pub fn source(&self) -> Square {
-        self.source_idx().try_into().unwrap()
-    }
-
-    /// Gets the target index from the compact move representation
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::notation::*;
-    ///
-    /// // the move from d2 -> d4
-    ///
-    /// let m = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-    ///
-    /// assert_eq!(m.target_idx(), usize::from(D4));
-    /// ```
-    pub fn target_idx(&self) -> usize {
-        ((self.0 & TARGET_IDX_MASK) >> TARGET_IDX_SHIFT).into()
-    }
-
-    /// Gets the target index from the compact move representation
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::notation::*;
-    ///
-    /// // the move from d2 -> d4
-    ///
-    /// let m = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-    ///
-    /// assert_eq!(m.target(), D4);
-    /// ```
-    pub fn target(&self) -> Square {
-        self.target_idx().try_into().unwrap()
-    }
-
-    /// True if the move indicates a promotion
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::notation::*;
-    ///
-    /// // the move from d2 -> d4
-    /// let m1 = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-    /// let m2 = Move::new(D7, D8, MoveType::PROMOTION_QUEEN);
-    /// assert!(!m1.is_promotion());
-    /// assert!(m2.is_promotion());
-    /// ```
-    pub fn is_promotion(&self) -> bool {
-        self.move_metadata().is_promotion()
-    }
-
-    /// Calculates the promotion piece is there is a promotion to be done.
-    ///
-    /// NOTE: Will return garbage for non-promotion moves. No checking is done ahead of time.
-    ///
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::constants::*;
-    /// # use hazel::notation::*;
-    /// # use hazel::types::Piece;
-    /// // the move from d2 -> d4
-    /// let m1 = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-    /// let m2 = Move::new(D7, D8, MoveType::PROMOTION_QUEEN);
-    /// // assert!(m1.promotion_piece()); DON'T DO THIS! It's not a promotion so this is misinterpreting the union type.
-    /// assert_eq!(m2.promotion_piece(), Piece::Queen);
-    /// ```
-    pub fn promotion_piece(&self) -> Piece {
-        self.move_metadata().promotion_piece().unwrap()
-    }
-
-    /// Interprets the metadata bits when the piece is not a promotion. Use the provided `is_` functions
-    /// on MoveType to interpret the data.
-    /// ```
-    /// # use hazel::coup::rep::*;
-    /// # use hazel::notation::*;
-    /// // the move from d2 -> d4
-    /// let m1 = Move::new(D2, D4, MoveType::QUIET);
-    /// assert!(m1.move_metadata().is_quiet());
-    /// ```
-    pub fn move_metadata(&self) -> MoveType {
-        MoveType::new(self.0 & METADATA_MASK)
-    }
 
     #[instrument(skip(context))]
     pub fn compile<C>(&self, context: &C) -> Vec<Alteration> where C : Query {
@@ -386,6 +222,10 @@ impl Move {
         let target_occupant = context.get(target);
 
         let contextprime = self.disambiguate(context).unwrap();
+
+        if let Some(metadata) = context.try_metadata() {
+            
+        }
 
 
         let alterations = match contextprime {
@@ -502,7 +342,6 @@ impl Move {
     }
 
     // Some proxy methods
-    // TODO: Maybe move metadata to a simple enum we can just match on, would make the #make/unmake implementations nicer
     #[inline(always)]
     pub fn is_capture(&self) -> bool {
         self.move_metadata().is_capture()
@@ -557,333 +396,6 @@ impl Move {
         match color {
             Color::WHITE => self.source() == E1 && self.target() == C1,
             Color::BLACK => self.source() == E8 && self.target() == C8,
-        }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    mod creation {
-        use super::*;
-
-        #[test]
-        fn new_move() {
-            let m = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-            assert_eq!(m.source(), D2);
-            assert_eq!(m.target(), D4);
-            assert!(!m.is_promotion());
-            assert!(!m.is_null());
-            assert!(m.is_double_pawn_push_for(Color::WHITE));
-        }
-
-        #[test]
-        fn empty_move() {
-            let m = Move::empty();
-            assert_eq!(m.0, 0);
-        }
-
-        #[test]
-        fn null_move() {
-            let m = Move::null();
-            assert!(m.is_null());
-        }
-
-    }
-
-    mod from_notation {
-        use super::*;
-
-        #[test]
-        fn quiet_move_parses_correctly() {
-            let m = Move::from_notation("d2", "d4", MoveType::QUIET);
-
-            assert_eq!(m.source_idx(), 0o13);
-            assert_eq!(m.target_idx(), 0o33);
-            assert!(!m.is_promotion());
-            assert!(m.move_metadata().is_quiet());
-        }
-
-        #[test]
-        fn promotion_move_parses_correctly() {
-            let pm = Move::from_notation("d7", "d8", MoveType::PROMOTION_QUEEN);
-            assert_eq!(pm.source_idx(), 0o63);
-            assert_eq!(pm.target_idx(), 0o73);
-            assert!(pm.is_promotion());
-            assert_eq!(pm.promotion_piece(), Piece::Queen)
-        }
-    }
-
-    mod castling {
-        use super::*;
-
-        mod white {
-            use super::*;
-
-            #[test]
-            fn short_castle_parses_correctly() {
-                let m = Move::short_castle(Color::WHITE);
-                assert_eq!(m.source(), E1);
-                assert_eq!(m.target(), G1);
-                assert!(m.is_short_castle());
-            }
-
-            #[test]
-            fn long_castle_parses_correctly() {
-                let m = Move::long_castle(Color::WHITE);
-                assert_eq!(m.source(), E1);
-                assert_eq!(m.target(), C1);
-                assert!(m.is_long_castle());
-            }
-        }
-
-        mod black {
-            use super::*;
-
-            #[test]
-            fn short_castle_parses_correctly() {
-                let m = Move::short_castle(Color::BLACK);
-                assert_eq!(m.source(), E8);
-                assert_eq!(m.target(), G8);
-                assert!(m.is_short_castle());
-            }
-
-            #[test]
-            fn long_castle_parses_correctly() {
-                let m = Move::long_castle(Color::BLACK);
-                assert_eq!(m.source(), E8);
-                assert_eq!(m.target(), C8);
-                assert!(m.is_long_castle());
-            }
-        }
-    }
-
-    mod disambiguate {
-        use crate::Alter;
-        use crate::board::PieceBoard;
-
-        use super::*;
-
-        #[test]
-        fn quiet_move_disambiguates_correctly() {
-            let m = Move::new(D2, D3, MoveType::UCI_AMBIGUOUS);
-            let mut context = PieceBoard::default();
-            context.set_startpos();
-
-            assert_eq!(m.disambiguate(&context).unwrap(), MoveType::QUIET);
-        }
-
-        #[test]
-        fn capture_move_disambiguates_correctly() {
-            let m = Move::new(C3, D4, MoveType::UCI_AMBIGUOUS);
-            let mut context = PieceBoard::default();
-            context.alter_mut(Alteration::place(C3, Occupant::pawn(Color::WHITE)));
-            context.alter_mut(Alteration::place(D4, Occupant::pawn(Color::BLACK)));
-
-            assert_eq!(m.disambiguate(&context).unwrap(), MoveType::CAPTURE);
-        }
-
-        #[test]
-        fn double_pawn_move_disambiguates_correctly() {
-            let m = Move::new(D2, D4, MoveType::UCI_AMBIGUOUS);
-            let mut context = PieceBoard::default();
-            context.set_startpos();
-
-            assert_eq!(m.disambiguate(&context).unwrap(), MoveType::DOUBLE_PAWN);
-        }
-
-        #[test]
-        fn short_castle_disambiguates_correctly() {
-            let m = Move::new(E1, G1, MoveType::UCI_AMBIGUOUS);
-            let mut context = PieceBoard::default();
-            context.set_startpos();
-            context.alter_mut(Alteration::remove(F1, Occupant::bishop(Color::WHITE)));
-            context.alter_mut(Alteration::remove(G1, Occupant::knight(Color::WHITE)));
-
-            assert_eq!(m.disambiguate(&context).unwrap(), MoveType::SHORT_CASTLE);
-        }
-
-        #[test]
-        fn long_castle_disambiguates_correctly() {
-            let m = Move::new(E1, C1, MoveType::UCI_AMBIGUOUS);
-            let mut context = PieceBoard::default();
-            context.set_startpos();
-            context.alter_mut(Alteration::remove(B1, Occupant::knight(Color::WHITE)));
-            context.alter_mut(Alteration::remove(C1, Occupant::bishop(Color::WHITE)));
-
-            assert_eq!(m.disambiguate(&context).unwrap(), MoveType::LONG_CASTLE);
-        }
-    }
-
-    mod to_star {
-        use crate::board::PieceBoard;
-        use crate::Alter;
-
-        use super::*;
-
-        mod pgn {
-            use super::*;
-
-            #[test]
-            fn to_pgn_quiet_move() {
-                let m = Move::new(D2, D3, MoveType::QUIET);
-                let mut context = PieceBoard::default();
-                context.set_startpos();
-
-                assert_eq!(m.to_pgn(&context), "d3");
-            }
-
-            #[test]
-            fn to_pgn_capture_move() {
-                let m = Move::new(D4, E5, MoveType::CAPTURE);
-                let mut context = PieceBoard::default();
-                context.alter_mut(Alteration::place(D4, Occupant::white_pawn()));
-                context.alter_mut(Alteration::place(E5, Occupant::black_pawn()));
-
-                assert_eq!(m.to_pgn(&context), "dxe5");
-            }
-
-            #[test]
-            fn to_pgn_double_pawn_move() {
-                let m = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-                let mut context = PieceBoard::default();
-                context.set_startpos();
-
-                assert_eq!(m.to_pgn(&context), "d4");
-            }
-
-            #[test]
-            fn to_pgn_short_castle() {
-                let m = Move::short_castle(Color::WHITE);
-                let mut context = PieceBoard::default();
-                // NOTE: Context does not mean it checks to see if the move is legal, in this position
-                // white cannot castle, eppur si muove.
-                context.set_startpos();
-
-                assert_eq!(m.to_pgn(&context), "O-O");
-            }
-
-            #[test]
-            fn to_pgn_long_castle() {
-                let m = Move::long_castle(Color::WHITE);
-                let mut context = PieceBoard::default();
-                // NOTE: same caveat as short castling.
-                context.set_startpos();
-
-                assert_eq!(m.to_pgn(&context), "O-O-O");
-            }
-
-            #[test]
-            fn to_pgn_promotion_move() {
-                let m = Move::new(D7, D8, MoveType::PROMOTION_QUEEN);
-                let mut context = PieceBoard::default();
-                context.alter_mut(Alteration::place(D7, Occupant::white_pawn()));
-
-                assert_eq!(m.to_pgn(&context), "d8=Q");
-            }
-
-            #[test]
-            fn to_pgn_capture_promotion_move() {
-                let m = Move::new(C7, D8, MoveType::PROMOTION_CAPTURE_QUEEN);
-                let mut context = PieceBoard::default();
-                context.alter_mut(Alteration::place(C7, Occupant::white_pawn()));
-                context.alter_mut(Alteration::place(D8, Occupant::black_pawn()));
-
-                assert_eq!(m.to_pgn(&context), "cxd8=Q");
-            }
-        }
-
-        mod uci {
-            use super::*;
-
-            #[test]
-            fn to_uci_quiet_move() {
-                let m = Move::new(D2, D3, MoveType::QUIET);
-                assert_eq!(m.to_uci(), "d2d3");
-            }
-
-            #[test]
-            fn to_uci_capture_move() {
-                let m = Move::new(D4, E5, MoveType::CAPTURE);
-                assert_eq!(m.to_uci(), "d4e5");
-            }
-
-            #[test]
-            fn to_uci_double_pawn_move() {
-                let m = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-                assert_eq!(m.to_uci(), "d2d4");
-            }
-
-            #[test]
-            fn to_uci_short_castle() {
-                let m = Move::short_castle(Color::WHITE);
-                assert_eq!(m.to_uci(), "e1g1");
-            }
-
-            #[test]
-            fn to_uci_long_castle() {
-                let m = Move::long_castle(Color::WHITE);
-                assert_eq!(m.to_uci(), "e1c1");
-            }
-
-            #[test]
-            fn to_uci_promotion_move() {
-                let m = Move::new(D7, D8, MoveType::PROMOTION_QUEEN);
-                assert_eq!(m.to_uci(), "d7d8q");
-            }
-
-            #[test]
-            fn to_uci_capture_promotion_move() {
-                let m = Move::new(C7, D8, MoveType::PROMOTION_CAPTURE_QUEEN);
-                assert_eq!(m.to_uci(), "c7d8q");
-            }
-        }
-    }
-
-    mod proxy_methods {
-        use super::*;
-
-        #[test]
-        fn is_capture() {
-            let m = Move::new(D2, D4, MoveType::CAPTURE);
-            assert!(m.is_capture());
-        }
-
-        #[test]
-        fn is_short_castle() {
-            let m = Move::short_castle(Color::WHITE);
-            assert!(m.is_short_castle());
-        }
-
-        #[test]
-        fn is_long_castle() {
-            let m = Move::long_castle(Color::WHITE);
-            assert!(m.is_long_castle());
-        }
-
-        #[test]
-        fn is_en_passant() {
-            let m = Move::new(D6, E7, MoveType::EP_CAPTURE);
-            assert!(m.is_en_passant());
-        }
-
-        #[test]
-        fn is_double_pawn_push_for() {
-            let m = Move::new(D2, D4, MoveType::DOUBLE_PAWN);
-            assert!(m.is_double_pawn_push_for(Color::WHITE));
-        }
-
-        #[test]
-        fn is_short_castling_move_for() {
-            let m = Move::short_castle(Color::WHITE);
-            assert!(m.is_short_castling_move_for(Color::WHITE));
-        }
-
-        #[test]
-        fn is_long_castling_move_for() {
-            let m = Move::long_castle(Color::WHITE);
-            assert!(m.is_long_castling_move_for(Color::WHITE));
         }
     }
 }
